# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepriseAI Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/referral/generate-code": {
      "post": {
        "tags": ["referral"],
        "summary": "Generate Referral Code",
        "description": "Generate a new referral code for the logged-in user.\nOne code per user (previous code becomes inactive if exists).",
        "operationId": "generate_referral_code_referral_generate_code_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralCodeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/referral/my-code": {
      "get": {
        "tags": ["referral"],
        "summary": "Get My Referral Code",
        "description": "Get the current active referral code for the logged-in user.",
        "operationId": "get_my_referral_code_referral_my_code_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ReferralCodeOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get My Referral Code Referral My Code Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/referral/info": {
      "get": {
        "tags": ["referral"],
        "summary": "Get My Referral Info",
        "description": "Get referral information for the logged-in user.\nIncludes points and active referral code.",
        "operationId": "get_my_referral_info_referral_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserReferralInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/referral/validate": {
      "post": {
        "tags": ["referral"],
        "summary": "Validate Code",
        "description": "Validate a referral code without redeeming it.\nUsed during signup form validation.",
        "operationId": "validate_code_referral_validate_post",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/referral/redeem": {
      "post": {
        "tags": ["referral"],
        "summary": "Redeem Code",
        "description": "Redeem a referral code for the current user.\nShould be called after signup to award points.",
        "operationId": "redeem_code_referral_redeem_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralRedeemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReferralCodeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": ["id", "code", "created_at", "expires_at", "is_active"],
        "title": "ReferralCodeOut"
      },
      "ReferralRedeemRequest": {
        "properties": {
          "referral_code": {
            "type": "string",
            "title": "Referral Code"
          }
        },
        "type": "object",
        "required": ["referral_code"],
        "title": "ReferralRedeemRequest"
      },
...
      "ReferralValidationResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "referrer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referrer Name"
          }
        },
        "type": "object",
        "required": ["valid", "message"],
        "title": "ReferralValidationResponse"
      },
...
      "UserReferralInfo": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "referral_points": {
            "type": "integer",
            "title": "Referral Points"
          },
          "referral_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Code"
          },
          "referral_code_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Code Expires At"
          }
        },
        "type": "object",
        "required": ["user_id", "referral_points"],
        "title": "UserReferralInfo",
        "description": "Info about a user's referral points"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/admin/auth/login"
          }
        }
      }
    }
  }
}